Django/pyproject.toml
[tool.poetry]
name = "oz-django"
version = "0.1.0"
description = ""
authors = ["park2do <ejfhr7777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.3"
djangorestframework = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
